+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `impersonateUser` mutation
 ## [1.9.1] - 2021-11-22
 
 ### Fixed